[Refactor]: 패키지 명 수정 & 코드 리팩토링
