+++ domain/src/main/kotlin/ac/dnd/bookkeeping/android/domain/model/sociallogin/UserModel.kt
@@ -1,4 +1,4 @@
-package ac.dnd.bookkeeping.android.domain.model.social_login
+package ac.dnd.bookkeeping.android.domain.model.sociallogin
 
 class UserModel(
     val socialId: Long,
